+++ packages/server/lib/utils/utils.ts
@@ -1,6 +1,7 @@
 import { fileURLToPath } from 'node:url';
 import path from 'node:path';
 import type { Request } from 'express';
+import { serializeError } from 'serialize-error';
 import type { DBUser, Provider, ProviderJwt, ProviderTwoStep } from '@nangohq/types';
 import type { Result } from '@nangohq/utils';
 import { Err, Ok } from '@nangohq/utils';
@@ -268,4 +269,29 @@ export function getOrchestrator() {
 export function isBinaryContentType(contentType: string | undefined): boolean {
     if (!contentType) return false;
     return BINARY_CONTENT_TYPES.some((type) => contentType.startsWith(type));
+}
+
+export function stringifyEnrichedError(err: unknown, opts?: { pretty?: boolean; stack?: boolean }) {
+    const serialized = serializeError(err);
+    const allowedErrorProperties = ['name', 'message', 'provider_error_payload', ...(opts?.stack ? ['stack', 'cause'] : [])];
+
+    const enriched: Record<string, unknown> = {
+        ...serialized
+    };
+
+    // Handle Boom-style error objects like the ones use in simpleOauth2
+    // These errors often contain valuable debugging information in `data.payload` that
+    // wouldn't be included in standard error serialization. We explicitly extract this
+    // as 'provider_error_payload' since it typically contains API-specific error details
+    // that are useful for debugging.
+    if (typeof err === 'object' && err != null) {
+        const anyErr = err as any;
+
+        const payload = anyErr.data?.payload;
+        if (payload && typeof payload === 'object') {
+            enriched['provider_error_payload'] = payload;
+        }
+    }
+    const filtered: Record<string, unknown> = Object.fromEntries(Object.entries(enriched).filter(([key]) => allowedErrorProperties.includes(key)));
+    return JSON.stringify(filtered, null, opts?.pretty ? 2 : undefined);
 }
