+++ packages/server/package.json
@@ -59,6 +59,7 @@
         "rate-limiter-flexible": "5.0.3",
         "redis": "4.6.13",
         "semver": "7.6.3",
+        "serialize-error": "11.0.3",
         "simple-oauth2": "5.1.0",
         "ua-parser-js": "2.0.2",
         "uuid": "9.0.0",
