+++ packages/utils/lib/errors.ts
@@ -17,31 +17,6 @@ export function stringifyError(err: unknown, opts?: { pretty?: boolean; stack?: 
     return JSON.stringify(serializeError(err), ['name', 'message', ...(opts?.stack ? ['stack', 'cause'] : [])], opts?.pretty ? 2 : undefined);
 }
 
-export function stringifyEnrichedError(err: unknown, opts?: { pretty?: boolean; stack?: boolean }) {
-    const serialized = serializeError(err);
-    const allowedErrorProperties = ['name', 'message', 'provider_error_payload', ...(opts?.stack ? ['stack', 'cause'] : [])];
-
-    const enriched: Record<string, unknown> = {
-        ...serialized
-    };
-
-    // Handle Boom-style error objects like the ones use in simpleOauth2
-    // These errors often contain valuable debugging information in `data.payload` that
-    // wouldn't be included in standard error serialization. We explicitly extract this
-    // as 'provider_error_payload' since it typically contains API-specific error details
-    // that are useful for debugging.
-    if (typeof err === 'object' && err != null) {
-        const anyErr = err as any;
-
-        const payload = anyErr.data?.payload;
-        if (payload && typeof payload === 'object') {
-            enriched['provider_error_payload'] = payload;
-        }
-    }
-    const filtered: Record<string, unknown> = Object.fromEntries(Object.entries(enriched).filter(([key]) => allowedErrorProperties.includes(key)));
-    return JSON.stringify(filtered, null, opts?.pretty ? 2 : undefined);
-}
-
 export function initSentry({ dsn, hash, applicationName }: { dsn: string | undefined; hash?: string | undefined; applicationName: string }) {
     Sentry.init({
         dsn: dsn || '',
